MAJOR fix for the core Social Proof Token and indexer servers logic
"---
name: SPT pipeline verification
overview: Audit and test the social indexer → Postgres schema → readers → GraphQL/social-server path so SPT quantities stay full nano-SPT integers end-to-end, and confirm no stale `symbol`/`name` dependencies after Move `TokenInfo` changes.
todos:
  - id: grep-stale-spt-metadata
    content: Grep spt/social_proof_token codepaths for symbol/name or TokenPoolCreated legacy fields; fix any stragglers
    status: pending
  - id: verify-no-float-truncation
    content: Trace spt ingest + SQL views for float casts or divide-by-1e9 on token quantities
    status: pending
  - id: optional-nano-test
    content: "Add indexer test: odd nano amount round-trips through holding row"
    status: pending
  - id: run-crate-tests
    content: Run tests for indexer, reader, graphql, social-server; clippy on touched crates
    status: pending
isProject: false
---

# SPT pipeline audit (nano amounts + removed symbol/name)

## Current design (already aligned)

- **Storage:** [`spt_holdings.amount`](crates/myso-indexer-alt-social-schema/src/schema.rs) is `Int8` / `BIGINT`. Same for [`spt_pools.circulating_supply`](crates/myso-indexer-alt-social-schema/src/schema.rs), [`spt_transactions.amount`](crates/myso-indexer-alt-social-schema/src/schema.rs), etc. These are **fixed-point integers** (nano-SPT), not floats. Comments: migration [`20260428120000_spt_nano_spt_column_comments/up.sql`](crates/myso-indexer-alt-social-schema/migrations/20260428120000_spt_nano_spt_column_comments/up.sql) and model docs in [`models/spt.rs`](crates/myso-indexer-alt-social-schema/src/models/spt.rs) (`SPT_AMOUNT_NANO_SCALE`, `SptHoldingRow.amount`).
- **Ingestion:** [`handlers/spt.rs`](crates/myso-indexer-alt-social/src/handlers/spt.rs) maps chain `u64` amounts via `require_chain_u64_as_i64` / `json_opt_chain_u64_as_i64_for_db` with **no scaling** (no divide-by-10^9). Values that do not fit `i64` skip the batch and increment metrics (`record_spt_u64_amount_exceeds_i64`).
- **Launch distribution:** [`proportional_spt_launch_holdings`](crates/myso-indexer-alt-social/src/handlers/spt_handler.rs) uses `i128` integer math, floors per reserver, assigns **remainder to owner**, and asserts the sum equals `circulating_supply` — preserving total nano-SPT.
- **social-server JSON:** [`reader/types/spt.rs`](crates/myso-social-server/src/reader/types/spt.rs) uses `json_string_i64::serialize` on `circulating_supply`, transaction `amount`, and holding `balance` so JSON clients do not lose precision for large integers.
- **GraphQL:** [`SptHolding`](crates/myso-indexer-alt-graphql/schema.graphql) exposes `amount: BigInt!` (nano-SPT). [`api/types/spt.rs`](crates/myso-indexer-alt-graphql/src/api/types/spt.rs) builds `BigInt` from reader `balance`. [`SocialProofToken`](crates/myso-indexer-alt-graphql/schema.graphql) has **no** `symbol`/`name` fields; [`SocialProofTokenInfo`](crates/myso-indexer-alt-social-reader/src/profile.rs) likewise omits them.
- **Removed on-chain fields:** `spt_pools` has no symbol/name columns; event parsing uses [`BcsTokenPoolCreatedEvent`](crates/myso-indexer-alt-social/src/handlers/events.rs) without those fields.

The `symbol` appearing in GraphQL schema snapshots (e.g. near line 773) is on **coin/treasury types**, not SPT.

## Verification checklist (read + targeted tests)

1. **Stray references:** Repo-wide search for `spt` + `symbol` / `name` in Rust/SQL/GraphQL outside unrelated domains (platforms, profiles, coins). Confirm zero expectations on `TokenPoolCreatedEvent` JSON keys `symbol`/`name` in tests.
2. **Event → DB path:** Trace one `TokenBoughtEvent` / `TokenSoldEvent` / `TokenPoolCreatedEvent` flow: [`events.rs` `parse_spt_event`](crates/myso-indexer-alt-social/src/handlers/events.rs) → [`spt.rs` row builders](crates/myso-indexer-alt-social/src/handlers/spt.rs) → Diesel inserts. Confirm no `as f64`, `/ 1_000_000_000` on SPT **token** amounts (only display math in clients if any).
3. **Aggregation SQL:** Grep views and analytics queries that `SUM(spt_holdings.amount)` or join holdings; confirm they stay in `bigint` and are not cast to `float`/`real`.
4. **Optional strengthen test:** Add (or extend) a unit test in [`spt.rs` tests](crates/myso-indexer-alt-social/src/handlers/spt.rs) with a **non–multiple-of-10^9** nano amount (e.g. `123456789`) on buy/sell and assert the inserted holding delta matches exactly — proves no accidental rounding to “whole tokens.”
5. **CI commands:** Run `cargo test -p myso-indexer-alt-social --lib`, `-p myso-indexer-alt-social-reader`, `-p myso-indexer-alt-graphql` (schema snapshots if touched), `-p myso-social-server`, and `cargo check` / `cargo xclippy` on touched crates.

## Edge case to document (not a “nano vs integer” bug)

- **i64 ceiling:** If chain amounts ever exceed `i64::MAX`, rows are skipped rather than truncated silently — acceptable unless product requires `NUMERIC`/unsigned; call out in runbooks if needed.

No schema migration is required for nano precision **if** the pipeline already stores raw chain units in `BIGINT`; the plan is **verification + optional test**, not a redesign.
"

+++ crates/myso-indexer-alt-graphql/src/api/types/profile.rs
@@ -647,12 +647,12 @@ impl SocialProofToken {
         }
     }
 
-    async fn total_reserved(&self) -> i64 {
-        self.inner.total_reserved
+    async fn total_reserved(&self) -> BigInt {
+        BigInt::from(self.inner.total_reserved)
     }
 
-    async fn required_threshold(&self) -> i64 {
-        self.inner.required_threshold
+    async fn required_threshold(&self) -> BigInt {
+        BigInt::from(self.inner.required_threshold)
     }
 
     /// Circulating supply in nano-SPT (`10^9` units per display token).
